Crate ethcore_network

Modules

client_version

Parse ethereum client ID strings and provide querying functionality

Structs

Error

The Error type.

IpFilter

NetworkConfiguration

Network service configuration

PeerCapabilityInfo

SessionCapabilityInfo

SessionInfo

Shared session information

Enums

AllowIP

IP fiter

ConnectionDirection

Filtered connection direction.

DisconnectReason

ErrorKind

The kind of an error.

NetworkIoMessage

Messages used to communitate with the event loop from other threads.

NonReservedPeerMode

Non-reserved peer modes.

Traits

ConnectionFilter

Connection filter. Each connection is checked against connection_allowed.

NetworkContext

IO access point. This is passed to all IO handlers and provides an interface to the IO subsystem.

NetworkProtocolHandler

Network IO protocol handler. This needs to be implemented for each new subprotocol. All the handler function are called from within IO event loop. Message is the type for message data.

Type Definitions

NodeId

Node public key

PacketId

Protocol handler level packet id

PeerId

Local (temporary) peer session ID.

ProtocolId

Protocol / handler id

TimerToken

Timer ID