Module ethcore_network::client_version

Parse ethereum client ID strings and provide querying functionality

Structs

ParityClientData

Description of the software version running in a peer according to https://github.com/ethereum/wiki/wiki/Client-Version-Strings This structure as it is represents the format used by Parity clients. Other vendors may provide additional fields.

Enums

ClientVersion

Enum describing the version of the software running on a peer.

Traits

ClientCapabilities

Provide information about what a particular version of a peer software can do