Enum ethcore_network::NetworkIoMessage

pub enum NetworkIoMessage {
    AddHandler {
        handler: Arc<NetworkProtocolHandler + Sync>,
        protocol: ProtocolId,
        versions: Vec<u8>,
        packet_count: u8,
    },
    AddTimer {
        protocol: ProtocolId,
        token: TimerToken,
        delay: u64,
    },
    InitPublicInterface,
    Disconnect(PeerId),
    DisablePeer(PeerId),
    NetworkStarted(String),
}

Messages used to communitate with the event loop from other threads.

Variants
AddHandler

Register a new protocol handler.

Fields

handler: Arc<NetworkProtocolHandler + Sync>

Handler shared instance.

protocol: ProtocolId

Protocol Id.

versions: Vec<u8>

Supported protocol versions.

packet_count: u8

Number of packet IDs reserved by the protocol.

AddTimer

Register a new protocol timer

Fields

protocol: ProtocolId

Protocol Id.

token: TimerToken

Timer token.

delay: u64

Timer delay in milliseconds.

InitPublicInterface

Initliaze public interface.

Disconnect(PeerId)

Disconnect a peer.

DisablePeer(PeerId)

Disconnect and temporary disable peer.

NetworkStarted(String)

Network has been started with the host as the given enode.